Move tests in a different package to be sure we test the public API
