+++ java/src/test/java/io/vertx/ext/eventbus/client/test/JacksonCodecTest.java
@@ -1,4 +1,4 @@
-package io.vertx.ext.eventbus.client;
+package io.vertx.ext.eventbus.client.test;
 
 import io.vertx.ext.eventbus.client.json.JacksonCodec;
